Avoid erase existing certificates

+++ kube/kubelet/scripts/retrieve-certificates.sh
@@ -12,18 +12,24 @@ do
     sleep 1
 done
 # Root CA
-curl -s ${PIDALIO_URL}/certs/ca\?token\=${PIDALIO_TOKEN} > ca.json
-cat ca.json | jq -r .cert > /etc/kubernetes/ssl/ca.pem
+if [ ! -e /etc/kubernetes/ssl/ca.pem ]; then
+    curl -s ${PIDALIO_URL}/certs/ca\?token\=${PIDALIO_TOKEN} > ca.json
+    cat ca.json | jq -r .cert > /etc/kubernetes/ssl/ca.pem
+fi
 if [[ "${MASTER}" == "true" ]]
 then
   # Server Certificate
-  curl -s ${PIDALIO_URL}/certs/server\?token\=${PIDALIO_TOKEN}\&ip=${NODE_PUBLIC_IP} > server.json
-  cat server.json | jq -r .privateKey > /etc/kubernetes/ssl/server-key.pem
-  cat server.json | jq -r .cert > /etc/kubernetes/ssl/server.pem
+  if [ ! -e /etc/kubernetes/ssl/server.pem ]; then
+      curl -s ${PIDALIO_URL}/certs/server\?token\=${PIDALIO_TOKEN}\&ip=${NODE_PUBLIC_IP} > server.json
+      cat server.json | jq -r .privateKey > /etc/kubernetes/ssl/server-key.pem
+      cat server.json | jq -r .cert > /etc/kubernetes/ssl/server.pem
+  fi
 fi
-curl -s ${PIDALIO_URL}/certs/node\?token\=${PIDALIO_TOKEN}\&fqdn=${NODE_FQDN}\&ip=${NODE_PUBLIC_IP} > node.json
-cat node.json | jq -r .privateKey > /etc/kubernetes/ssl/node-key.pem
-cat node.json | jq -r .cert > /etc/kubernetes/ssl/node.pem
+if [ ! -e /etc/kubernetes/ssl/node.pem ]; then
+    curl -s ${PIDALIO_URL}/certs/node\?token\=${PIDALIO_TOKEN}\&fqdn=${NODE_FQDN}\&ip=${NODE_PUBLIC_IP} > node.json
+    cat node.json | jq -r .privateKey > /etc/kubernetes/ssl/node-key.pem
+    cat node.json | jq -r .cert > /etc/kubernetes/ssl/node.pem
+fi
 cat <<EOF > /etc/kubernetes/kubeconfig.yaml
 apiVersion: v1
 kind: Config
